add UserInGroup to TeachersSubjectsController

+++ HomewOurK.WebAPI/Controllers/TeachersSubjectsController.cs
@@ -1,4 +1,6 @@
 ﻿using HomewOurK.Application.Interfaces;
+using HomewOurK.Infrastructure.Services;
+using HomewOurK.WebAPI.Helpers;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -10,26 +12,42 @@ namespace HomewOurK.WebAPI.Controllers
 	public class TeachersSubjectsController : ControllerBase
 	{
 		private readonly ITeacherService _teacherService;
+		private readonly IHttpContextAccessor _httpContextAccessor;
+		private readonly IUserService _userService;
 
-		public TeachersSubjectsController(ITeacherService teacherService)
+		public TeachersSubjectsController(ITeacherService teacherService, IHttpContextAccessor httpContextAccessor, IUserService userService)
 		{
 			_teacherService = teacherService;
+			_httpContextAccessor = httpContextAccessor;
+			_userService = userService;
 		}
 
 		[HttpPost]
 		public IActionResult AddSubjectToTeacher(int teacherId, int groupId, int subjectId)
 		{
-			if (_teacherService.AddSubject(teacherId, groupId, subjectId))
-				return Ok();
-			return BadRequest();
+			var email = CookieHelper.GetEmailByCookie(_httpContextAccessor);
+
+			if (_userService.UserInGroup(groupId, email))
+			{
+				if (_teacherService.AddSubject(teacherId, groupId, subjectId))
+					return Ok();
+				return BadRequest();
+			}
+			return Unauthorized();
 		}
 
 		[HttpDelete]
 		public IActionResult DeleteSubjectFromTeacher(int teacherId, int groupId, int subjectId)
 		{
-			if (_teacherService.DeleteSubject(teacherId, groupId, subjectId))
-				return Ok();
-			return BadRequest();
+			var email = CookieHelper.GetEmailByCookie(_httpContextAccessor);
+
+			if (_userService.UserInGroup(groupId, email))
+			{
+				if (_teacherService.DeleteSubject(teacherId, groupId, subjectId))
+					return Ok();
+				return BadRequest();
+			}
+			return Unauthorized();
 		}
 	}
 }
